Issue #28 clean-up and rename bookmarks component and adjust routing
