+++ src/application/__tests__/routing.test.ts
@@ -39,8 +39,8 @@ describe('the routePathDefinition function', () => {
         expect(routePathDefinition(Routes.RecommendedTopics)).toEqual('/recommended-topics');
     });
 
-    it('For Routes.BookmarkedTopics returns path: "/bookmarked"', () => {
-        expect(routePathDefinition(Routes.Bookmarked)).toEqual('/bookmarked');
+    it('For Routes.BookmarkedTopics returns path: "/bookmarks"', () => {
+        expect(routePathDefinition(Routes.Bookmarks)).toEqual('/bookmarks');
     });
 
     it('For Routes.Search returns path: "/search"', () => {
