+++ src/components/main/main_page_switcher.tsx
@@ -9,7 +9,7 @@ import { TaskDetailConnectedComponent } from '../topics/task_detail_connected_co
 import { ServiceListConnectedComponent } from '../services/service_list_connected_component';
 import { QuestionnaireConnectedComponent } from '../questionnaire/questionnaire_connected_component';
 import { RecommendedTopicsConnectedComponent } from '../recommended_topics/recommended_topics_connected_component';
-import { BookmarkedTopicsConnectedComponent } from '../bookmarked_topics/bookmarked_topics_connected_component';
+import { BookmarksConnectedComponent } from '../bookmarks/bookmarks_connected_component';
 import { SearchConnectedComponent } from '../search/search_connected_component';
 import { Routes, routePathDefinition, routePathWithoutParameter } from '../../application/routing';
 import { Locale } from '../../locale';
@@ -31,7 +31,7 @@ export const MainPageSwitcherComponent: React.StatelessComponent<Props> = (props
         <Route exact path={routePathDefinition(Routes.TopicDetail)} component={TaskDetailConnectedComponent} />
         <Route exact path={routePathDefinition(Routes.Services)} component={ServiceListConnectedComponent} />
         <Route exact path={routePathDefinition(Routes.RecommendedTopics)} component={RecommendedTopicsConnectedComponent} />
-        <Route exact path={routePathDefinition(Routes.Bookmarked)} component={BookmarkedTopicsConnectedComponent} />
+        <Route exact path={routePathDefinition(Routes.Bookmarks)} component={BookmarksConnectedComponent} />
         <Route exact path={routePathDefinition(Routes.Search)} component={SearchConnectedComponent} />
         <Redirect to={defaultPath(props)} />
     </Switch>
